Introduce format version v2
