+++ code/go/internal/validator/spec_test.go
@@ -5,8 +5,9 @@
 package validator
 
 import (
-	"github.com/elastic/package-spec/code/go/internal/fspath"
 	"testing"
+
+	"github.com/elastic/package-spec/code/go/internal/fspath"
 
 	"github.com/Masterminds/semver/v3"
 	"github.com/stretchr/testify/require"
@@ -40,6 +41,7 @@ func TestNoBetaFeatures_Package_GA(t *testing.T) {
 	s := Spec{
 		*semver.MustParse("1.0.0"),
 		fspath.DirFS("testdata/fakespec"),
+		nil,
 		"1",
 	}
 	pkg, err := NewPackage("testdata/packages/features_ga")
@@ -54,12 +56,13 @@ func TestBetaFeatures_Package_GA(t *testing.T) {
 	s := Spec{
 		*semver.MustParse("1.0.0"),
 		fspath.DirFS("testdata/fakespec"),
+		nil,
 		"1",
 	}
 	pkg, err := NewPackage("testdata/packages/features_beta")
 	require.NoError(t, err)
 
 	errs := s.ValidatePackage(*pkg)
-	require.Len(t,errs, 1)
+	require.Len(t, errs, 1)
 	require.Equal(t, errs[0].Error(), "spec for [testdata/packages/features_beta/beta] defines beta features which can't be enabled for packages with a stable semantic version")
 }
