+++ spec.go
@@ -10,6 +10,7 @@ import (
 )
 
 //go:embed versions/1 versions/1/integration/_dev versions/1/integration/data_stream/_dev versions/1/input
+//go:embed versions/2 versions/2/integration/_dev versions/2/integration/data_stream/_dev versions/2/input
 var content embed.FS
 
 // FS returns an io/fs.FS for accessing the "package-spec/version" contents.
