+++ test/packages/good/manifest.yml
@@ -1,4 +1,4 @@
-format_version: 1.0.4
+format_version: 2.0.0
 name: good
 title: Good package
 description: This package is good.
