+++ versions/2/input/manifest.spec.yml
@@ -1,0 +1,121 @@
+##
+## Describes the specification for the input package's main manifest.yml file
+##
+spec:
+  # Everything under here follows JSON schema (https://json-schema.org/), written as YAML for readability
+  type: object
+  additionalProperties: false
+  properties:
+    format_version:
+      description: The version of the package specification format used by this package.
+      $ref: "../integration/manifest.spec.yml#/definitions/version"
+    name:
+      description: The name of the package.
+      type: string
+      pattern: '^[a-z0-9_]+$'
+      examples:
+      - apache
+    title:
+      description: The title of the package.
+      type: string
+      examples:
+      - Apache
+    description:
+      description: A longer description of the package.
+      type: string
+      examples:
+      - Apache Integration
+    version:
+      description: The version of the package.
+      $ref: "../integration/manifest.spec.yml#/definitions/version"
+    source:
+      $ref: "../integration/manifest.spec.yml#/definitions/source"
+    license:
+      description: The license under which the package is being released (deprecated, use subscription instead).
+      deprecated: true
+      type: string
+      enum:
+      - basic
+      default: basic
+      examples:
+      - basic
+    type:
+      description: The type of package.
+      type: string
+      enum:
+      - input
+      examples:
+      - input
+    categories:
+      $ref: "../integration/manifest.spec.yml#/definitions/categories"
+    conditions:
+      $ref: "../integration/manifest.spec.yml#/definitions/conditions"
+    policy_templates:
+      description: List of policy templates offered by this package.
+      type: array
+      items:
+        type: object
+        additionalProperties: false
+        properties:
+          name:
+            description: Name of policy template.
+            type: string
+            examples:
+              - apache
+          type:
+            description: Type of data stream
+            type: string
+            enum:
+              - metrics
+              - logs
+              - synthetics
+              - traces
+            examples:
+              - metrics
+          title:
+            description: Title of policy template.
+            type: string
+            examples:
+              - Apache logs and metrics
+          description:
+            description: Longer description of policy template.
+            type: string
+            examples:
+              - Collect logs and metrics from Apache instances
+          icons:
+            $ref: "../integration/manifest.spec.yml#/definitions/icons"
+          screenshots:
+            $ref: "../integration/manifest.spec.yml#/definitions/screenshots"
+          vars:
+            $ref: "../integration/data_stream/manifest.spec.yml#/definitions/vars"
+          input:
+            type: string
+            examples:
+              - aws/metrics
+              - s3
+              - file
+          template_path:
+            description: "Path to Elasticsearch index template for stream."
+            type: string
+        required:
+          - name
+          - title
+          - description
+          - type
+          - input
+    icons:
+      $ref: "../integration/manifest.spec.yml#/definitions/icons"
+    screenshots:
+      $ref: "../integration/manifest.spec.yml#/definitions/screenshots"
+    vars:
+      $ref: "../integration/data_stream/manifest.spec.yml#/definitions/vars"
+    owner:
+      $ref: "../integration/manifest.spec.yml#/definitions/owner"
+  required:
+  - format_version
+  - name
+  - title
+  - description
+  - version
+  - type
+  - owner
