+++ versions/2/integration/_dev/deploy/agent/custom-agent.spec.yml
@@ -1,0 +1,30 @@
+##
+## Describes the specification for variations available when deploying a package's
+## integration service(s).
+##
+spec:
+  # Everything under here follows JSON schema (https://json-schema.org/), written as YAML for readability
+  type: object
+  additionalProperties: false
+  properties:
+    version:
+      description: Docker Compose version
+      type: string
+      pattern: '^2.3$'
+    services:
+      description: Service list
+      type: object
+      additionalProperties: false
+      properties:
+        docker-custom-agent:
+          description: Custom agent service definition
+          type: object
+          not:
+            anyOf:
+              - required:
+                  - hostname
+      required:
+        - docker-custom-agent
+  required:
+    - version
+    - services
