+++ versions/2/integration/_dev/deploy/agent/spec.yml
@@ -1,0 +1,9 @@
+spec:
+  additionalContents: false
+  contents:
+    - description: Custom agent definition.
+      type: file
+      name: "custom-agent.yml"
+      contentMediaType: "application/x-yaml"
+      required: true
+      $ref: "./custom-agent.spec.yml"
