+++ versions/2/integration/_dev/deploy/docker/spec.yml
@@ -1,0 +1,8 @@
+spec:
+  additionalContents: true
+  contents:
+    - description: Docker Compose configuration file for deploying package's integration service(s)
+      type: file
+      name: "docker-compose.yml"
+      contentMediaType: "application/x-yaml"
+      required: true
