+++ versions/2/integration/_dev/deploy/spec.yml
@@ -1,0 +1,25 @@
+spec:
+  additionalContents: false
+  contents:
+    - description: Folder containing deployment files for a Docker-based service deployment
+      type: folder
+      name: docker
+      $ref: "./docker/spec.yml"
+    - description: Folder containing deployment files for a Terraform-based service deployment
+      type: folder
+      name: tf
+      $ref: "./tf/spec.yml"
+    - description: Folder containing deployment files for a Cusom Agent service deployment
+      type: folder
+      name: agent
+      $ref: "./agent/spec.yml"
+    - description: Folder containing deployment files for a Kubernetes service deployment
+      type: folder
+      name: k8s
+      $ref: "./k8s/spec.yml"
+    - description: Configuration file describing variations available for deployment
+      type: file
+      name: "variants.yml"
+      contentMediaType: "application/x-yaml"
+      required: false
+      $ref: "./variants.spec.yml"
