+++ versions/2/integration/_dev/deploy/tf/env.spec.yml
@@ -1,0 +1,36 @@
+##
+## Describes the specification for variations available when deploying a package's
+## integration service(s).
+##
+spec:
+  # Everything under here follows JSON schema (https://json-schema.org/), written as YAML for readability
+  type: object
+  additionalProperties: false
+  properties:
+    version:
+      description: Docker Compose version
+      type: string
+      pattern: '^2.3$'
+    services:
+      description: Service list
+      type: object
+      additionalProperties: false
+      properties:
+        terraform:
+          description: Terraform service definition
+          type: object
+          additionalProperties: false
+          properties:
+            environment:
+              description: List of environment variables
+              type: array
+              items:
+                type: string
+                pattern: '^[A-Za-z0-9_]+=.+$'
+          required:
+            - environment
+      required:
+        - terraform
+  required:
+    - version
+    - services
