+++ versions/2/integration/data_stream/_dev/test/pipeline/common_config.spec.yml
@@ -1,0 +1,20 @@
+spec:
+  type: object
+  additionalProperties: false
+  properties:
+    skip:
+      $ref: '../skip.spec.yml#/definitions/skip'
+    fields:
+      description: Field definitions
+      type: object
+      additionalProperties: true
+    dynamic_fields:
+      description: Dynamic fields with regular expressions defining their variable values.
+      type: object
+    multiline:
+      description: Multi-line configuration
+      type: object
+      additionalProperties: true
+    numeric_keyword_fields:
+      description: List of keyword type fields allowed to have a numeric value.
+      type: array
