+++ versions/2/integration/data_stream/_dev/test/static/config.spec.yml
@@ -1,0 +1,10 @@
+##
+## Describes the specification for a system test configuration file
+##
+spec:
+  # Everything under here follows JSON schema (https://json-schema.org/), written as YAML for readability
+  type: object
+  additionalProperties: true
+  properties:
+    skip:
+      $ref: "./../skip.spec.yml#/definitions/skip"
