+++ versions/2/integration/data_stream/_dev/test/system/config.spec.yml
@@ -1,0 +1,14 @@
+##
+## Describes the specification for a system test configuration file
+##
+spec:
+  # Everything under here follows JSON schema (https://json-schema.org/), written as YAML for readability
+  type: object
+  additionalProperties: true
+  properties:
+    skip:
+      $ref: "./../skip.spec.yml#/definitions/skip"
+    wait_for_data_timeout:
+      description: Timeout for waiting for metrics data during a system test.
+      type: string
+      example: 10m
