+++ versions/2/integration/data_stream/_dev/test/system/spec.yml
@@ -1,0 +1,10 @@
+spec:
+  additionalContents: false
+  contents:
+    - description: Configuration of test case in yaml format.
+      type: file
+      pattern: '^test-[a-z0-9_.-]+-config.yml$'
+      contentMediaType: "application/x-yaml"
+      required: false
+      $ref: "./config.spec.yml"
+
