+++ versions/2/integration/data_stream/spec.yml
@@ -1,0 +1,80 @@
+spec:
+  additionalContents: false
+  totalContentsLimit: 500
+  contents:
+  - description: Folder containing a single data stream definition
+    type: folder
+    pattern: '^([a-z0-9]{2}|[a-z0-9][a-z0-9_]+[a-z0-9])$'
+    required: true
+    additionalContents: false
+    contents:
+    - description: A data stream's manifest file
+      type: file
+      contentMediaType: "application/x-yaml"
+      sizeLimit: 5MB
+      name: "manifest.yml"
+      required: true
+      $ref: "./manifest.spec.yml"
+    - description: Folder containing field definitions
+      type: folder
+      name: fields
+      required: true
+      $ref: "./fields/spec.yml"
+    - description: Folder containing agent-related definitions
+      type: folder
+      name: agent
+      required: false
+      additionalContents: false
+      $ref: "./agent/spec.yml"
+    - description: Folder containing Elasticsearch assets
+      type: folder
+      name: elasticsearch
+      additionalContents: false
+      contents:
+      - description: Folder containing Elasticsearch ILM Policy Definition
+        type: folder
+        name: ilm
+        additionalContents: false
+        contents:
+        - description: Supporting ILM policy definitions in YAML
+          type: file
+          pattern: '^.+\.yml$'
+          # TODO Determine if special handling of `---` is required (issue: https://github.com/elastic/package-spec/pull/54)
+          contentMediaType: "application/x-yaml; require-document-dashes=true"
+          required: false
+        - description: Supporting ILM policy definitions in JSON
+          type: file
+          pattern: '^.+\.json$'
+          contentMediaType: "application/json"
+          required: false
+      - description: Folder containing Elasticsearch Ingest Node pipeline definitions
+        type: folder
+        name: ingest_pipeline
+        additionalContents: false
+        contents:
+        - description: Supporting ingest pipeline definitions in YAML
+          type: file
+          pattern: '^.+\.yml$'
+          # TODO Determine if special handling of `---` is required (issue: https://github.com/elastic/package-spec/pull/54)
+          contentMediaType: "application/x-yaml; require-document-dashes=true"
+          required: false
+        - description: Supporting ingest pipeline definitions in JSON
+          type: file
+          pattern: '^.+\.json$'
+          contentMediaType: "application/json"
+          required: false
+    - description: Sample event file
+      type: file
+      name: "sample_event.json"
+      contentMediaType: "application/json"
+      required: false
+    - description: Folder containing testing related files and sub-folders
+      type: folder
+      name: "test"
+      required: false
+    - description: Folder containing development resources
+      type: folder
+      name: _dev
+      required: false
+      visibility: private
+      $ref: "./_dev/spec.yml"
