+++ versions/2/integration/docs/spec.yml
@@ -1,0 +1,13 @@
+  spec:
+    additionalContents: false
+    contents:
+    - description: Main README file
+      type: file
+      contentMediaType: "text/markdown"
+      name: "README.md"
+      required: true
+    - description: Other README files (can be used by policy templates)
+      type: file
+      contentMediaType: "text/markdown"
+      pattern: '^.+.md'
+      required: false
