+++ versions/2/integration/elasticsearch/spec.yml
@@ -1,0 +1,36 @@
+spec:
+  additionalContents: false
+  contents:
+    - description: Folder containing ML model assets
+      type: folder
+      name: ml_model
+      required: false
+      contents:
+        - description: An ML model file
+          type: file
+          contentMediaType: "application/json"
+          pattern: '^{PACKAGE_NAME}_.+\.json$'
+    - description: Folder containing Elasticsearch Ingest Node pipeline definitions
+      type: folder
+      name: ingest_pipeline
+      additionalContents: true
+      contents:
+        - description: Supporting ingest pipeline definitions in YAML
+          type: file
+          sizeLimit: 3MB
+          pattern: '^.+\.yml$'
+          # TODO Determine if special handling of `---` is required (issue: https://github.com/elastic/package-spec/pull/54)
+          contentMediaType: "application/x-yaml; require-document-dashes=true"
+          required: false
+        - description: Supporting ingest pipeline definitions in JSON
+          type: file
+          sizeLimit: 3MB
+          pattern: '^.+\.json$'
+          contentMediaType: "application/json"
+          required: false
+    - description: Folder containing Elasticsearch Transforms
+      # https://www.elastic.co/guide/en/elasticsearch/reference/current/transforms.html
+      type: folder
+      name: transform
+      required: false
+      $ref: "./transform/spec.yml"
