+++ versions/2/integration/elasticsearch/transform/manifest.spec.yml
@@ -1,0 +1,18 @@
+##
+## Describes the specification for a transform's manifest.yml file
+##
+spec:
+  type: object
+  additionalProperties: false
+  definitions:
+    index_template:
+      $ref: "../../data_stream/manifest.spec.yml#/properties/elasticsearch/properties/index_template"
+  properties:
+    destination_index_template:
+      description: Elasticsearch index template for the transform's destination index
+      $ref: "#/definitions/index_template"
+      required: false
+    start:
+      description: Determines if the transform will be started upon installation
+      type: boolean
+      default: true
