+++ versions/2/integration/spec.yml
@@ -1,0 +1,68 @@
+##
+## Entrypoint of "integration packages" specification.
+##
+## Describes the folders and files that make up a package.
+##
+spec:
+  additionalContents: true
+  totalContentsLimit: 65535
+  totalSizeLimit: 250MB
+  sizeLimit: 150MB
+  configurationSizeLimit: 5MB
+  relativePathSizeLimit: 3MB
+  fieldsPerDataStreamLimit: 2048
+  contents:
+  - description: The main package manifest file
+    type: file
+    contentMediaType: "application/x-yaml"
+    sizeLimit: 5MB
+    name: "manifest.yml"
+    required: true
+    $ref: "./manifest.spec.yml"
+  - description: The package's CHANGELOG file
+    type: file
+    contentMediaType: "application/x-yaml"
+    name: "changelog.yml"
+    required: true
+    $ref: "./changelog.spec.yml"
+  - description: The package's NOTICE file
+    type: file
+    contentMediaType: "text/plain"
+    name: "NOTICE.txt"
+    required: false
+  - description: The package's license file
+    type: file
+    contentMediaType: "text/plain"
+    name: "LICENSE.txt"
+    required: false
+  - description: Folder containing data stream definitions
+    type: folder
+    name: data_stream
+    required: false
+    $ref: "./data_stream/spec.yml"
+  - description: Folder containing documentation for the package
+    type: folder
+    name: docs
+    required: true
+    $ref: "./docs/spec.yml"
+  - description: Folder containing agent-related definitions
+    type: folder
+    name: agent
+    required: false
+    $ref: "./agent/spec.yml"
+  - description: Folder containing Kibana assets used by the package
+    type: folder
+    name: kibana
+    required: false
+    $ref: "./kibana/spec.yml"
+  - description: Folder containing development resources
+    type: folder
+    name: _dev
+    required: false
+    visibility: private
+    $ref: "./_dev/spec.yml"
+  - description: Folder containing Elasticsearch assets used by the package
+    type: folder
+    name: elasticsearch
+    required: false
+    $ref: "./elasticsearch/spec.yml"
